Fix KDBX4 key calculation on 32 bit platforms (#45)
* Fix #41: Fix incorrect key calculation on 32 bit platforms

* add 32bit test to ci

+++ src/result.rs
@@ -27,7 +27,7 @@ pub enum DatabaseIntegrityError {
     },
     HeaderHashMismatch,
     BlockHashMismatch {
-        block_index: usize,
+        block_index: u64,
     },
     InvalidKDBXIdentifier,
     InvalidKDBXVersion {
